+++ test/contracts/mock/StorageGetter.sol
@@ -14,7 +14,11 @@ struct PackedStruct {
 contract StorageGetter {
   address internal _address;
   uint256 internal _constructorUint256;
+  int56 internal _int56;
+  int256 internal _int256;
   uint256 internal _uint256;
+  bytes internal _bytes;
+  bytes32 internal _bytes32;
   bool internal _bool;
   SimpleStruct internal _simpleStruct;
   PackedStruct internal _packedStruct;
@@ -23,6 +27,7 @@ contract StorageGetter {
   mapping(bytes5 => bool) _bytes5ToBoolMap;
   mapping(address => bool) _addressToBoolMap;
   mapping(address => address) _addressToAddressMap;
+  uint256[] internal _uint256Array;
 
   // Testing storage slot packing.
   bool internal _packedA;
@@ -36,8 +41,24 @@ contract StorageGetter {
     return _constructorUint256;
   }
 
+  function getInt56() public view returns (int56 _out) {
+    return _int56;
+  }
+
+  function getInt256() public view returns (int256 _out) {
+    return _int256;
+  }
+
   function getUint256() public view returns (uint256 _out) {
     return _uint256;
+  }
+
+  function getBytes() public view returns (bytes memory _out) {
+    return _bytes;
+  }
+
+  function getBytes32() public view returns (bytes32 _out) {
+    return _bytes32;
   }
 
   function setPackedA(bool _val) external {
@@ -64,6 +85,10 @@ contract StorageGetter {
     return _simpleStruct;
   }
 
+  function getPackedStruct() public view returns (PackedStruct memory _out) {
+    return _packedStruct;
+  }
+
   function getUint256MapValue(uint256 _key) public view returns (uint256 _out) {
     return _uint256Map[_key];
   }
@@ -82,6 +107,10 @@ contract StorageGetter {
 
   function getAddressToAddressMapValue(address _key) public view returns (address _out) {
     return _addressToAddressMap[_key];
+  }
+
+  function getUint256Array() public view returns (uint256[] memory _out) {
+    return _uint256Array;
   }
 
   function getPackedAddress() public view returns (address) {